chore: move server side libraries to lib/server
