+++ src/routes/api/ideas/delete/+server.js
@@ -1,7 +1,7 @@
 import { error } from '@sveltejs/kit';
 import { ObjectId } from 'mongodb';
-import { getDb } from '$lib/mongodb-client';
-import { getPayloadFromJWTCookie } from '$lib/utils';
+import { getDb } from '$lib/server/mongodb-client';
+import { getPayloadFromJWTCookie } from '$lib/server/utils';
 
 import dotenv from 'dotenv';
 dotenv.config();
